+++ framework/src/Volo.Abp.Ldap.Abstractions/Volo/Abp/Ldap/Localization/en-GB.json
