+++ framework/src/Volo.Abp.Ldap.Abstractions/Volo/Abp/Ldap/Localization/pt-BR.json
