+++ framework/src/Volo.Abp.Ldap.Abstractions/Volo/Abp/Ldap/Localization/ro-RO.json
